requirements/dev.txt
-r common.txt

django-debug-toolbar==2.0